update users page to use live data
