+++ src/app/admin/(admin)/users/_active-users-card.tsx
@@ -1,0 +1,15 @@
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
+
+export default async function ActiveUsersCard({ activeUsers }: { activeUsers: number }) {
+    return (
+        <Card>
+            <CardHeader>
+                <CardTitle>Active Users</CardTitle>
+                <CardDescription>Currently active accounts</CardDescription>
+            </CardHeader>
+            <CardContent>
+                <p className="text-3xl font-bold">{activeUsers}</p>
+            </CardContent>
+        </Card>
+    )
+}
